Modules for API and Utility Functions are created
