+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "liveServer.settings.CustomBrowser": "firefox",
+    "liveServer.settings.port": 5501
+}
