+++ index.html
@@ -9,11 +9,11 @@
   <body>
     <div class="container">
       <p class="comment-loader">Пожалуйста, подождите, комментарии загружаются...</p>
-      <ul id="comment-list" class="comments">jjjj</ul>
+      <ul id="comment-list" class="comments"></ul>
       <div id="input-form" class="add-form"></div>
       <button id="remove-comment-button" class="remove-comment-button">Удалить последний комментарий</button>
     </div>
   </body>
 
-  <script src="hw.js"></script>
+  <script type="module" src="main.js"></script>
 </html>
